requirements.txt
boto
awscli 
langchainhub
langchain_community
bs4
tiktoken
openai
chromadb
streamlit
pypdf
faiss-cpu
-e.